requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.5
Werkzeug==3.0.1
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
Flask-Limiter==3.5.0
marshmallow==3.20.1
python-socketio==5.10.0
python-engineio==4.8.0
psycopg2-binary==2.9.9
Flask-Admin==1.6.1
Pillow==11.0.0
Flask-Mail==0.9.1
gunicorn==21.2.0
eventlet==0.35.2